Automatic Process Service/slower/requirements.txt
psycopg2-binary
confluent_kafka